Move more PostgreSQL-specific functionality out of base dialect: user-defined types / domains, standard confirming strings setting, server type
