+++ api/src/org/labkey/api/data/dialect/BasePostgreSqlDialect.java
@@ -21,14 +21,12 @@ import org.apache.logging.log4j.Level;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.labkey.api.collections.CaseInsensitiveMapWrapper;
-import org.labkey.api.collections.CopyOnWriteHashMap;
 import org.labkey.api.collections.CsvSet;
 import org.labkey.api.collections.Sets;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.ConnectionWrapper;
 import org.labkey.api.data.ConnectionWrapper.Closer;
 import org.labkey.api.data.DatabaseIdentifier;
-import org.labkey.api.data.DbSchema;
 import org.labkey.api.data.DbScope;
 import org.labkey.api.data.DbScope.LabKeyDataSource;
 import org.labkey.api.data.ExceptionFramework;
@@ -37,16 +35,13 @@ import org.labkey.api.data.MetadataSqlSelector;
 import org.labkey.api.data.PropertyStorageSpec;
 import org.labkey.api.data.RuntimeSQLException;
 import org.labkey.api.data.SQLFragment;
-import org.labkey.api.data.Selector;
 import org.labkey.api.data.SqlExecutingSelector.ConnectionFactory;
 import org.labkey.api.data.SqlExecutor;
-import org.labkey.api.data.SqlSelector;
 import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.dialect.LimitRowsSqlGenerator.LimitRowsCustomizer;
 import org.labkey.api.data.dialect.LimitRowsSqlGenerator.StandardLimitRowsCustomizer;
 import org.labkey.api.exp.PropertyType;
-import org.labkey.api.util.ExceptionUtil;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.StringUtilsLabKey;
 import org.labkey.api.view.template.Warnings;
@@ -78,53 +73,10 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
     public static final String POSTGRES_LOCKS_TABLE_NAME = "pg_locks";
     public static final String POSTGRES_TABLE_SIZES_TABLE_NAME = "pg_tablesizes";
 
-    private final Map<String, Integer> _domainScaleMap = new CopyOnWriteHashMap<>();
-
     private HtmlString _adminWarning = null;
 
     // Default to 9 and let newer versions be refreshed later
     private int _majorVersion = 9;
-
-    // Specifies if this PostgreSQL server treats backslashes in string literals as normal characters (as per the SQL
-    // standard) or as escape characters (old, non-standard behavior). As of PostgreSQL 9.1, the setting
-    // standard_conforming_strings is on by default; before 9.1, it was off by default. We check the server setting
-    // when we prepare a new DbScope and use this when we escape and parse string literals.
-    private Boolean _standardConformingStrings = Boolean.TRUE;
-    private PostgreSqlServerType _serverType = PostgreSqlServerType.PostgreSQL;
-
-    public boolean getStandardConformingStrings()
-    {
-        // This should always be set before prior to getting the value
-        if (_standardConformingStrings == null)
-            throw new IllegalStateException("Standard Conforming Strings are not set for " + this);
-        return _standardConformingStrings;
-    }
-
-
-    @Override
-    protected DialectStringHandler createStringHandler()
-    {
-        // TODO: Should we look at "backslash_quote" setting instead/in addition?
-        if (getStandardConformingStrings())
-            return super.createStringHandler();
-        else
-            return new BackslashEscapingStringHandler();
-    }
-
-    public void setStandardConformingStrings(boolean standardConformingStrings)
-    {
-        _standardConformingStrings = standardConformingStrings;
-    }
-
-    public PostgreSqlServerType getServerType()
-    {
-        return _serverType;
-    }
-
-    public void setServerType(PostgreSqlServerType serverType)
-    {
-        _serverType = serverType;
-    }
 
     @Override
     protected @NotNull Set<String> getReservedWords()
@@ -630,66 +582,8 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
     }
 
     @Override
-    public String prepare(DbScope scope)
-    {
-        initializeUserDefinedTypes(scope);
-        determineSettings(scope);
-        return super.prepare(scope);
-    }
-
-    @Override
     public void prepareConnection(Connection conn)
     {
-    }
-
-    // When a new PostgreSQL DbScope is created, we enumerate the domains (user-defined types) in the public schema
-    // of the datasource, determine their "scale," and stash that information in a map associated with the DbScope.
-    // When the PostgreSQLColumnMetaDataReader reads metadata, it returns these scale values for all domains.
-    private void initializeUserDefinedTypes(DbScope scope)
-    {
-        // Skip domains query if connecting to LabKey Server - it has no user-defined types
-        if (getServerType().supportsSpecialMetadataQueries())
-        {
-            Selector selector = new SqlSelector(scope, "SELECT * FROM information_schema.domains");
-            selector.forEach(rs -> {
-                String schemaName = rs.getString("domain_schema");
-                String domainName = rs.getString("domain_name");
-                String dataType = rs.getString("data_type");
-                int scale;
-
-                if (dataType.startsWith("character"))
-                {
-                    String maxLength = rs.getString("character_maximum_length");
-
-                    // VARCHAR with no specific size has null maxLength... but character_octet_length seems okay
-                    scale = Integer.parseInt(null != maxLength ? maxLength : rs.getString("character_octet_length"));
-                }
-                else
-                {
-                    // Assume everything else is an integer for now. We should support more types for better external schema handling.
-                    scale = 4;
-                }
-
-                String key = getDomainKey(schemaName, domainName);
-                _domainScaleMap.put(key, scale);
-            });
-        }
-    }
-
-    private String getDomainKey(String schemaName, String domainName)
-    {
-        // Domain names are returned from column metadata fully qualified and quoted, so save them that way. See #26149.
-        return ("public".equals(schemaName) ? domainName : "\"" + schemaName + "\".\"" + domainName + "\"");
-    }
-
-    // Query any settings that may affect dialect behavior. Right now, only "standard_conforming_strings".
-    protected void determineSettings(DbScope scope)
-    {
-        if (getServerType().supportsSpecialMetadataQueries())
-        {
-            Selector selector = new SqlSelector(scope, "SELECT setting FROM pg_settings WHERE name = 'standard_conforming_strings'");
-            _standardConformingStrings = "on".equalsIgnoreCase(selector.getObject(String.class));
-        }
     }
 
     /**
@@ -854,7 +748,6 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
     @Override
     public ColumnMetaDataReader getColumnMetaDataReader(ResultSet rsCols, TableInfo table)
     {
-        // Retrieve and pass in the previously queried scale values for this scope.
         return new PostgreSqlColumnMetaDataReader(rsCols, table);
     }
 
@@ -1006,9 +899,9 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
                 .append(") AS TEXT) ~ '^[+-]?([0-9]+([.][0-9]*)?|[.][0-9]+)$' THEN 1 ELSE 0 END)");
     }
 
-    private class PostgreSqlColumnMetaDataReader extends ColumnMetaDataReader
+    public static class PostgreSqlColumnMetaDataReader extends ColumnMetaDataReader
     {
-        private final TableInfo _table;
+        protected final TableInfo _table;
 
         public PostgreSqlColumnMetaDataReader(ResultSet rsCols, TableInfo table)
         {
@@ -1046,39 +939,6 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
                 return _rsCols.getInt("SOURCE_DATA_TYPE");
             else
                 return sqlType;
-        }
-
-        @Override
-        public int getScale() throws SQLException
-        {
-            int sqlType = super.getSqlType();
-
-            return Types.DISTINCT == sqlType ? getDomainScale(getSqlTypeName()) : super.getScale();
-        }
-
-        private int getDomainScale(String domainName) throws SQLException
-        {
-            Integer scale = _domainScaleMap.get(domainName);
-
-            if (null == scale)
-            {
-                // Some domain wasn't there when we initialized the datasource, so reload now. This will happen at bootstrap.
-                DbSchema schema = _table.getSchema();
-                initializeUserDefinedTypes(schema.getScope());
-                scale = _domainScaleMap.get(domainName);
-
-                // If scale is still null, then we have a problem. We've seen occasional exception reports showing this,
-                // but haven't had the information to track it down... so log additional info.
-                if (null == scale)
-                {
-                    String message = "Null scale for \"" + domainName + "\" in column \"" + _table.getName() + "." + getName() + "\" in schema \"" + schema.getName() + "\"";
-                    ExceptionUtil.logExceptionToMothership(null, new Exception(message));
-                    assert false : message;
-                    return 4;   // Return something on production servers so schema can continue to load
-                }
-            }
-
-            return scale;
         }
 
         @Nullable
@@ -1290,20 +1150,6 @@ public abstract class BasePostgreSqlDialect extends SqlDialect
     {
         if (null != _adminWarning)
             warnings.add(_adminWarning);
-    }
-
-    @Override
-    public boolean isProcedureExists(DbScope scope, String schema, String name)
-    {
-        // Don't bother querying LabKey for stored procedures
-        return getServerType().supportsSpecialMetadataQueries() && super.isProcedureExists(scope, schema, name);
-    }
-
-    @Override
-    public boolean shouldTest()
-    {
-        // Don't test a LabKey data source
-        return getServerType().shouldTest();
     }
 
     @Override
