+++ core/src/org/labkey/core/dialect/PostgreSqlServerType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.api.data.dialect;
+package org.labkey.core.dialect;
 
 import java.util.Map;
